feat: Issue #751 - データ構造の標準化と一元化
